refactor(combobox): renamed combobox input as trigger
