+++ packages/components/combobox/src/ComboboxTrigger.styles.tsx
