+++ packages/components/combobox/src/index.ts
@@ -5,7 +5,6 @@ import { ComboboxProvider, useComboboxContext } from './ComboboxContext'
 import { Divider } from './ComboboxDivider'
 import { Empty } from './ComboboxEmpty'
 import { Group } from './ComboboxGroup'
-import { Input } from './ComboboxInput'
 import { Item } from './ComboboxItem'
 import { ItemIndicator } from './ComboboxItemIndicator'
 import { Items } from './ComboboxItems'
@@ -13,6 +12,7 @@ import { ItemText } from './ComboboxItemText'
 import { Label } from './ComboboxLabel'
 import { LeadingIcon } from './ComboboxLeadingIcon'
 import { Popover } from './ComboboxPopover'
+import { Trigger } from './ComboboxTrigger'
 
 export { useComboboxContext, ComboboxProvider }
 
@@ -25,7 +25,7 @@ export const Combobox: FC<ComboboxProps> & {
   Label: typeof Label
   Popover: typeof Popover
   Divider: typeof Divider
-  Input: typeof Input
+  Trigger: typeof Trigger
   LeadingIcon: typeof LeadingIcon
   Empty: typeof Empty
 } = Object.assign(Root, {
@@ -37,7 +37,7 @@ export const Combobox: FC<ComboboxProps> & {
   Label,
   Popover,
   Divider,
-  Input,
+  Trigger,
   LeadingIcon,
   Empty,
 })
@@ -51,6 +51,6 @@ ItemIndicator.displayName = 'Combobox.ItemIndicator'
 Label.displayName = 'Combobox.Label'
 Popover.displayName = 'Combobox.Popover'
 Divider.displayName = 'Combobox.Divider'
-Input.displayName = 'Combobox.Input'
+Trigger.displayName = 'Combobox.Trigger'
 LeadingIcon.displayName = 'Combobox.LeadingIcon'
 Empty.displayName = 'Combobox.Empty'
