+++ packages/components/combobox/src/tests/Combobox.test.tsx
@@ -10,7 +10,7 @@ describe('Combobox', () => {
   it('should render input and list of items', () => {
     render(
       <Combobox>
-        <Combobox.Input aria-label="Book" />
+        <Combobox.Trigger aria-label="Book" />
         <Combobox.Popover>
           <Combobox.Items>
             <Combobox.Item value="book-1">War and Peace</Combobox.Item>
@@ -37,7 +37,7 @@ describe('Combobox', () => {
       // Given a close combobox (default state)
       render(
         <Combobox>
-          <Combobox.Input aria-label="Book" />
+          <Combobox.Trigger aria-label="Book" />
           <Combobox.Popover>
             <Combobox.Items>
               <Combobox.Item value="book-1">War and Peace</Combobox.Item>
@@ -71,7 +71,7 @@ describe('Combobox', () => {
       // Given a close combobox without a popover(default state)
       render(
         <Combobox>
-          <Combobox.Input aria-label="Book" />
+          <Combobox.Trigger aria-label="Book" />
 
           <Combobox.Items>
             <Combobox.Item value="book-1">War and Peace</Combobox.Item>
@@ -104,7 +104,7 @@ describe('Combobox', () => {
       // Given a combobox that should remain opened
       render(
         <Combobox open>
-          <Combobox.Input aria-label="Book" />
+          <Combobox.Trigger aria-label="Book" />
           <Combobox.Popover>
             <Combobox.Items>
               <Combobox.Item value="book-1">War and Peace</Combobox.Item>
@@ -130,7 +130,7 @@ describe('Combobox', () => {
       // Given a combobox that should remain opened
       render(
         <Combobox defaultOpen>
-          <Combobox.Input aria-label="Book" />
+          <Combobox.Trigger aria-label="Book" />
           <Combobox.Popover>
             <Combobox.Items>
               <Combobox.Item value="book-1">War and Peace</Combobox.Item>
@@ -156,7 +156,7 @@ describe('Combobox', () => {
       // Given a combobox with no selected value yet
       render(
         <Combobox autoFilter>
-          <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+          <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
           <Combobox.Popover>
             <Combobox.Items>
               <Combobox.Empty>No results found</Combobox.Empty>
@@ -186,7 +186,7 @@ describe('Combobox', () => {
       // Given a combobox with no selected value yet
       render(
         <Combobox>
-          <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+          <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
           <Combobox.Popover>
             <Combobox.Items>
               <Combobox.Item value="book-1">War and Peace</Combobox.Item>
@@ -218,7 +218,7 @@ describe('Combobox', () => {
       // Given a combobox with a default selected value
       render(
         <Combobox defaultValue="book-2">
-          <Combobox.Input aria-label="Book" />
+          <Combobox.Trigger aria-label="Book" />
           <Combobox.Popover>
             <Combobox.Items>
               <Combobox.Item value="book-1">War and Peace</Combobox.Item>
@@ -237,7 +237,7 @@ describe('Combobox', () => {
       // Given a combobox with a default selected value
       render(
         <Combobox defaultValue="book-2">
-          <Combobox.Input aria-label="Book" />
+          <Combobox.Trigger aria-label="Book" />
           <Combobox.Popover>
             <Combobox.Items>
               <Combobox.Item value="book-1">
@@ -271,7 +271,7 @@ describe('Combobox', () => {
 
         return (
           <Combobox value={value} onValueChange={setValue}>
-            <Combobox.Input aria-label="Book" value={inputValue} onValueChange={setInputValue} />
+            <Combobox.Trigger aria-label="Book" value={inputValue} onValueChange={setInputValue} />
             <Combobox.Popover>
               <Combobox.Items>
                 <Combobox.Item value="book-1">War and Peace</Combobox.Item>
@@ -303,7 +303,7 @@ describe('Combobox', () => {
       // Given a combobox with no selected value yet
       render(
         <Combobox autoFilter>
-          <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+          <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
           <Combobox.Popover>
             <Combobox.Items>
               <Combobox.Item value="book-1">War and Peace</Combobox.Item>
