+++ packages/components/combobox/src/tests/multipleSelection.test.tsx
@@ -13,7 +13,7 @@ describe('Combobox', () => {
       // Given a combobox with no selected value yet
       render(
         <Combobox multiple>
-          <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+          <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
           <Combobox.Popover>
             <Combobox.Items>
               <Combobox.Item value="book-1">War and Peace</Combobox.Item>
@@ -47,7 +47,7 @@ describe('Combobox', () => {
       // Given a combobox with no selected value yet
       render(
         <Combobox multiple>
-          <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+          <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
           <Combobox.Popover>
             <Combobox.Items>
               <Combobox.Item value="book-1">War and Peace</Combobox.Item>
@@ -82,7 +82,7 @@ describe('Combobox', () => {
       // Given a combobox with no selected value yet
       render(
         <Combobox multiple>
-          <Combobox.Input aria-label="Book" placeholder="Pick a book" />
+          <Combobox.Trigger aria-label="Book" placeholder="Pick a book" />
           <Combobox.Popover>
             <Combobox.Items>
               <Combobox.Item value="book-1">War and Peace</Combobox.Item>
